requirements.txt
libcgroup >= 3.2.0
PyYAML >= 6.0.2
psutil >= 5.5.1
#cpufreq >= 0.3.5
